project/django_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
python-decouple==3.8
pillow==10.0.1
pillow-heif==0.13.0
ffmpeg-python==0.2.0
opencv-python==4.8.1.78
openai==1.3.5
pytesseract==0.3.10
celery==5.3.4
redis==5.0.1
django-filter==23.3
whitenoise==6.6.0
gunicorn==21.2.0
numpy==1.26.4
openpyxl==3.1.5
reportlab==4.0.8
# Supabase dependencies
supabase==2.3.4
django-storages==1.14.2
boto3==1.34.34